added  added ustring - utf8 string (c++11 std::string cant be memmoved and so used in fast_vector and some other containers)  string&stream utils refactored to ssu:: namespace  Config class can have multiple sections  added EventsHandler & EventsHandlerTyped for subscribing & emiting events  added ChunkedBuffer, used in MultiPool  added Items2 - facade class on bytes array with TypeInfo (calls destructors ...)  added Bits - dynamic size bitset  Mask class reworked to custom full static bitset  BitScanner allowing fast bit looping  added ObjFunc for function dispatch (faster than std::function), allowing compile time optimization
 added custom profiling
 added F8x8 - simple fixed sized bitmap font
 added TestBench - for running tests and profiling them

conditionally some SDL2 dependant things (guarded by USE_SDL2 define):
  - SDL2TestBench for simple SDL2 window with debug overlay with profiling info
  - SDL2Text for drawing F8x8 texts

+++ include/types/containers/Items.h
@@ -18,7 +18,7 @@ namespace grynca {
         }
 
         u32 size() {
-            return bytes_->size()/sizeof(ItemType);
+            return u32(bytes_->size()/sizeof(ItemType));
         }
 
         bool empty() {
